summary_tables for grouped_df release candidate for 0.8.0
Prior to this change the summary_table for grouped_by data.frames worked
for dplyr version 0.7.8 and below.  Release candidate for dplyr version
0.8.0, however, broke the prior version of summary_table due to a change
in the attributes of grouped_df objects.

This change adds code to summary_table.grouped_df which will account for
the attributes of dplyr version 0.7.8 and the attributes for version
0.8.0.

Fix #67