use attr(*, "label") to set names of rgroup
Prior to this change only the variable name would be used to label the
row groups for `qsummary` and `summary_table`

This change looks to see if the attribute "label" has been set for a
variable in the data frame and uses that instead of the variable name.
If the attribute is NULL then the variable name will be used.  Examples
to follow.

re #74