replace tidyverse and magrittr use in pkg_check
Prior to this change the `pkg_check` function used tibbles and
dplyr::bind_rows when that was not needed.  Also, the %>% was used, but
not needed to be.

This change replaced tibble with data.frame and bind_rows with
do.call(rbind, ...)

This change also adds an example for pkg_check

re #96, #97